requirements.txt
# Manuscript parsing
pymupdf>=1.23.0
python-docx>=0.8.11

# Chunking
llama-index>=0.10.0

# Embeddings & vector store
openai>=1.0.0
numpy>=1.26.0

# Canon Pack generation
anthropic>=0.25.0

# Schema validation
pydantic>=2.0.0

# Config
python-dotenv>=1.0.0
